ajout formulaire ajoutProduit

+++ src/main/java/DAO/ProduitDAO.java
@@ -139,7 +139,7 @@ public class ProduitDAO {
 
         
         public void ajoutProduit(String nom,int reference, int fournisseur, int categorie,
-                                    String Quantite_Par_Unite, float Prix_unitaire, int Unites_en_Stock, 
+                                    String Quantite_Par_Unite, Double Prix_unitaire, int Unites_en_Stock, 
                                     int Unites_commandees, int Niveau_de_reappro, int Indisponible) throws SQLException {
             int num = 0;
             String sql1 = "SELECT max(REFERENCE) as REFERENCE FROM Produit";
@@ -161,7 +161,7 @@ public class ProduitDAO {
                 stmt.setInt(3,fournisseur);
                 stmt.setInt(4,categorie);
                 stmt.setString(5,Quantite_Par_Unite);
-                stmt.setFloat(6,Prix_unitaire);
+                stmt.setDouble(6,Prix_unitaire);
                 stmt.setInt(7,Unites_en_Stock);
                 stmt.setInt(8,Unites_commandees);
                 stmt.setInt(9,Niveau_de_reappro);
